pyproject.toml
[tool.poetry]
name = "my-movie-db"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-decouple = "^3.7"
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
